feat(redux): consolidate types
