+++ apps/web/store/initial-state.ts
@@ -1,0 +1,32 @@
+import type { ReduxState } from './types'
+import { EmptyResponse } from '@modtree/utils'
+
+export const baseInitialState: ReduxState = {
+  base: {
+    user: EmptyResponse.User,
+    degree: EmptyResponse.Degree,
+    graph: EmptyResponse.Graph,
+    modulesCondensed: {},
+  },
+  flow: {
+    selection: [],
+    nodes: [],
+    edges: [],
+  },
+  modal: {
+    contextMenuProps: {
+      top: 0,
+      left: 0,
+    },
+    showUserProfile: false,
+    showContextMenu: false,
+    showModuleModal: false,
+    showDebugModal: false,
+    modalModule: EmptyResponse.Module,
+  },
+  search: {
+    moduleCondensed: [],
+    module: [],
+    hasResults: false,
+  },
+}
